chore: move sanitize into utils
