+++ src/prime_rl/utils/utils.py
@@ -1,6 +1,7 @@
 import asyncio
 import functools
 import importlib
+import math
 import os
 import sys
 from collections import defaultdict
@@ -125,6 +126,31 @@ def format_num(num: float | int, precision: int = 2) -> str:
         return f"{sign}{num / 1e6:.{precision}f}M"
     else:
         return f"{sign}{num / 1e9:.{precision}f}B"
+
+
+def sanitize(obj: Any) -> tuple[Any, list[str]]:
+    """Recursively drop non-finite floats (NaN/inf), which are not valid JSON.
+    Returns the sanitized object and the dotted paths of the dropped values."""
+    dropped_paths: list[str] = []
+
+    def keep(item: Any, path: str) -> bool:
+        if isinstance(item, float) and not math.isfinite(item):
+            dropped_paths.append(path)
+            return False
+        return True
+
+    def walk(value: Any, path: str) -> Any:
+        if isinstance(value, dict):
+            return {
+                key: walk(item, child)
+                for key, item in value.items()
+                if keep(item, child := f"{path}.{key}" if path else key)
+            }
+        if isinstance(value, list):
+            return [walk(item, child) for index, item in enumerate(value) if keep(item, child := f"{path}[{index}]")]
+        return value
+
+    return walk(obj, ""), dropped_paths
 
 
 def get_free_port() -> int:
